Trying to cache GitHub releases data for later actions
